[REFACTOR] #36 - 연락처 리스트를 Manager 클래스로 분리
